refactor: radioState 파일명 오류 수정
